requirements.txt
python-dotenv~=0.21.0
dnspython~=2.3.0
mongoengine~=0.27.0
pymongo~=4.3.3
PyYAML==6.0
uvicorn~=0.23.1
fastapi~=0.100.0
openai
langchain
pydantic
redis
websockets
rq
bcrypt
PyJWT
requests
JinJa2
httpx
stripe
certifi
black